mlspark/requirements.txt
pymongo>=4.6.0
google-cloud-storage>=2.16.0
numpy==1.24.4
grpcio==1.58.0
